LibWeb: Split StyleValueList out of StyleValue.{h,cpp}
