+++ Userland/Libraries/LibWeb/CSS/StyleValues/StyleValueList.cpp
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2018-2020, Andreas Kling <kling@serenityos.org>
+ * Copyright (c) 2021, Tobias Christiansen <tobyase@serenityos.org>
+ * Copyright (c) 2021-2023, Sam Atkins <atkinssj@serenityos.org>
+ * Copyright (c) 2022-2023, MacDue <macdue@dueutil.tech>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#include "StyleValueList.h"
+
+namespace Web::CSS {
+
+bool StyleValueList::Properties::operator==(Properties const& other) const
+{
+    return separator == other.separator && values.span() == other.values.span();
+}
+
+ErrorOr<String> StyleValueList::to_string() const
+{
+    auto separator = ""sv;
+    switch (m_properties.separator) {
+    case Separator::Space:
+        separator = " "sv;
+        break;
+    case Separator::Comma:
+        separator = ", "sv;
+        break;
+    default:
+        VERIFY_NOT_REACHED();
+    }
+
+    StringBuilder builder;
+    for (size_t i = 0; i < m_properties.values.size(); ++i) {
+        TRY(builder.try_append(TRY(m_properties.values[i]->to_string())));
+        if (i != m_properties.values.size() - 1)
+            TRY(builder.try_append(separator));
+    }
+    return builder.to_string();
+}
+
+}
